removing old hbuild, removing dead code and renaming rebuild
