--- hbuild/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "hbuild"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-socket2 = "0.5"
-helixdb = { path = "../helixdb", features = ["server"] }
-aws-config = "1.1.1"
-aws-sdk-s3 = "1.11.0"
-tokio = { version = "1.44.2", features = ["full"] }
-anyhow = "1.0"
-serde = { version = "1.0", features = ["derive"] }
-sonic-rs = "0.5.0"
-futures = "0.3"
-
